Try and fix manifests
