+++ .github/workflows/test.yml
@@ -44,7 +44,7 @@ jobs:
 
   concurrency-checking:
     runs-on: ubuntu-latest
-    container: swift:5.8-jammy
+    container: swiftlang/swift:nightly-5.9-jammy@sha256:a3c3ec5e4436c14e44759e38416bb0d46813dc6dd050e787ef3a80b2c3051a36
     env:
       STRICT_CONCURRENCY: "true"
     steps:
